cherry-pick: AOT: add jsb interface type (#14368)
