+++ cocos/asset/assets/asset.jsb.ts
@@ -25,6 +25,7 @@
 import { cclegacy, js, _decorator, path, jsbUtils, CallbacksInvoker, applyMixins } from '../../core';
 import { getUrlWithUuid } from '../asset-manager/helper';
 import { patch_cc_Asset } from '../../native-binding/decorators';
+import type { Asset as JsbAsset } from './asset';
 
 declare const jsb: any;
 
@@ -114,9 +115,8 @@ assetProto.toString = function () {
 
 assetProto.createNode = null!;
 
-// @ts-ignore
-export type Asset = jsb.Asset;
-export const Asset = jsb.Asset;
+export type Asset = JsbAsset;
+export const Asset: typeof JsbAsset = jsb.Asset;
 
 cclegacy.Asset = jsb.Asset;
