+++ cocos/asset/assets/render-texture.jsb.ts
@@ -26,6 +26,7 @@ import { cclegacy } from '../../core';
 import { Filter, PixelFormat, WrapMode } from './asset-enum';
 import './asset';
 import { patch_cc_RenderTexture } from '../../native-binding/decorators';
+import type { RenderTexture as JsbRenderTexture } from './render-texture';
 
 declare const jsb: any;
 const renderTextureProto: any = jsb.RenderTexture.prototype;
@@ -33,9 +34,8 @@ const textureBaseProto: any = jsb.TextureBase.prototype;
 
 renderTextureProto.createNode = null!;
 
-// @ts-ignore
-export type RenderTexture = jsb.RenderTexture;
-export const RenderTexture: any = jsb.RenderTexture;
+export type RenderTexture = JsbRenderTexture;
+export const RenderTexture: typeof JsbRenderTexture = jsb.RenderTexture;
 
 RenderTexture.Filter = Filter;
 RenderTexture.PixelFormat = PixelFormat;
