+++ cocos/asset/assets/rendering-sub-mesh.jsb.ts
@@ -24,6 +24,9 @@
 
 import { Vec3 } from "../../core";
 import { Attribute, PrimitiveMode, Buffer } from "../../gfx";
+import type { RenderingSubMesh as JsbRenderingSubMesh } from './rendering-sub-mesh';
+
+declare const jsb: any;
 
 /**
  * @en Array views for index buffer
@@ -71,8 +74,8 @@ export interface IFlatBuffer {
     buffer: Uint8Array;
 }
 
-export type RenderingSubMesh = jsb.RenderingSubMesh;
-export const RenderingSubMesh = jsb.RenderingSubMesh;
+export type RenderingSubMesh = JsbRenderingSubMesh;
+export const RenderingSubMesh: typeof JsbRenderingSubMesh = jsb.RenderingSubMesh;
 
 const renderingSubMeshProto = RenderingSubMesh.prototype;
