+++ cocos/asset/assets/texture-2d.jsb.ts
@@ -21,6 +21,8 @@
  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  THE SOFTWARE.
 */
+
+import { EDITOR, TEST } from 'internal:constants'
 import { ImageAsset } from './image-asset';
 import { SimpleTexture } from './simple-texture';
 import { TextureBase } from './texture-base.jsb';
@@ -28,13 +30,15 @@ import { js, cclegacy } from '../../core';
 import { Filter, PixelFormat, WrapMode } from './asset-enum';
 import './simple-texture';
 import { patch_cc_Texture2D } from '../../native-binding/decorators';
+import type { Texture2D as JsbTexture2D } from './texture-2d';
 
+declare const jsb: any;
 const texture2DProto: any = jsb.Texture2D.prototype;
 
 texture2DProto.createNode = null!;
 
-export type Texture2D = jsb.Texture2D;
-export const Texture2D: any = jsb.Texture2D;
+export type Texture2D = JsbTexture2D;
+export const Texture2D: typeof JsbTexture2D = jsb.Texture2D;
 
 Texture2D.Filter = Filter;
 Texture2D.PixelFormat = PixelFormat;
