+++ cocos/asset/assets/texture-base.jsb.ts
@@ -21,12 +21,16 @@
  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  THE SOFTWARE.
 */
+
 import { TEST, EDITOR } from 'internal:constants';
 import { deviceManager } from '../../gfx';
 import { cclegacy } from '../../core';
 import { Filter, PixelFormat, WrapMode } from './asset-enum';
 import './asset';
 import { patch_cc_TextureBase } from '../../native-binding/decorators';
+import type { TextureBase as JsbTextureBase } from './texture-base';
+
+declare const jsb: any;
 
 const textureBaseProto: any = jsb.TextureBase.prototype;
 
@@ -78,8 +82,8 @@ textureBaseProto._getGFXPixelFormat = function (format) {
 
 textureBaseProto.createNode = null!;
 
-export type TextureBase = jsb.TextureBase;
-export const TextureBase: any = jsb.TextureBase;
+export type TextureBase = JsbTextureBase;
+export const TextureBase: typeof JsbTextureBase = jsb.TextureBase;
 
 TextureBase.Filter = Filter;
 TextureBase.PixelFormat = PixelFormat;
