+++ cocos/asset/assets/texture-cube.jsb.ts
@@ -26,6 +26,7 @@ import { js, cclegacy } from '../../core';
 import './simple-texture';
 import { EDITOR, TEST } from 'internal:constants';
 import { patch_cc_TextureCube } from '../../native-binding/decorators';
+import type { TextureCube as JsbTextureCube } from './texture-cube';
 
 const textureCubeProto: any = jsb.TextureCube.prototype;
 interface ITextureCubeSerializeData {
@@ -68,9 +69,8 @@ textureCubeProto.createNode = null!;
 
 declare const jsb: any;
 
-// @ts-expect-error jsb.TextureCube is exported from cpp
-export type TextureCube = jsb.TextureCube;
-export const TextureCube: any = jsb.TextureCube;
+export type TextureCube = JsbTextureCube;
+export const TextureCube: typeof JsbTextureCube = jsb.TextureCube;
 
 TextureCube.Filter = Filter;
 TextureCube.PixelFormat = PixelFormat;
