+++ cocos/render-scene/core/program-lib.ts
@@ -40,7 +40,7 @@ import { debug, cclegacy } from '../../core';
 
 const _dsLayoutInfo = new DescriptorSetLayoutInfo();
 
-interface IDefineRecord extends EffectAsset.IDefineInfo {
+export interface IDefineRecord extends EffectAsset.IDefineInfo {
     _map: (value: any) => number;
     _offset: number;
 }
@@ -315,7 +315,7 @@ function processShaderInfo (
  * @en The global maintainer of all shader resources.
  * @zh 维护 shader 资源实例的全局管理器。
  */
-class ProgramLib {
+export class ProgramLib {
     protected _templates: Record<string, IProgramInfo> = {}; // per shader
     protected _cache: Record<string, Shader> = {};
     protected _templateInfos: Record<number, ITemplateInfo> = {};
