+++ cocos/render-scene/core/render-scene.jsb.ts
@@ -23,6 +23,9 @@
 */
 
 import { Node } from '../../scene-graph';
+import { RenderScene as JsbRenderScene } from './render-scene';
+
+declare const jsb: any;
 
 export interface IRenderSceneInfo {
     name: string;
@@ -49,5 +52,5 @@ Object.defineProperty(renderSceneProto, 'mainLight', {
         return this.getMainLight();
     }
 });
-
-export const RenderScene = jsb.RenderScene;
+export const RenderScene: typeof JsbRenderScene = jsb.RenderScene;
+export type RenderScene = JsbRenderScene;
