+++ cocos/render-scene/core/render-window.jsb.ts
@@ -21,6 +21,9 @@
  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  THE SOFTWARE.
 */
+
+import type { RenderWindow as JsbRenderWindow } from './render-window';
+
 export interface IRenderWindowInfo {
     title?: string;
     width: number;
@@ -30,6 +33,5 @@ export interface IRenderWindowInfo {
 }
 
 declare const jsb: any;
-export const RenderWindow = jsb.RenderWindow;
-// @ts-ignore
-export type RenderWindow = jsb.RenderWindow;
+export const RenderWindow: typeof JsbRenderWindow = jsb.RenderWindow;
+export type RenderWindow = JsbRenderWindow;
