+++ cocos/render-scene/index.ts
@@ -34,7 +34,8 @@ export { addStage };
 export * from './core/constants';
 export * from './core/pass-utils';
 export * from './core/pass';
-export * from './core/program-lib';
+export type { IProgramInfo, ITemplateInfo } from './core/program-lib';
+export { programLib, getDeviceShaderVersion } from './core/program-lib';
 export * from './core/texture-buffer-pool';
 export * from './core/material-instance';
 export * from './core/pass-instance';
