+++ cocos/render-scene/scene/model.jsb.ts
@@ -24,6 +24,9 @@
 
 import { Attribute, deviceManager } from '../../gfx';
 import { Vec3 } from '../../core';
+import type { Model as JsbModel } from './model';
+
+declare const jsb: any;
 
 export interface IInstancedAttributeBlock {
     buffer: Uint8Array;
@@ -40,7 +43,8 @@ export enum ModelType {
     LINE,
 }
 
-export const Model = jsb.Model;
+export const Model: typeof JsbModel = jsb.Model;
+export type Model = JsbModel;
 
 const modelProto: any = Model.prototype;
