+++ cocos/root.jsb.ts
@@ -27,11 +27,13 @@ import { DataPoolManager } from './3d/skeletal-animation/data-pool-manager';
 import { Device, deviceManager } from './gfx';
 import { settings, Settings, warnID, Pool, macro } from './core';
 import { ForwardPipeline } from './rendering';
+import type { Root as JsbRoot } from './root';
 
 declare const nr: any;
 declare const jsb: any;
 
-export const Root = jsb.Root;
+export const Root: typeof JsbRoot = jsb.Root;
+export type Root = JsbRoot;
 
 enum LightType {
     DIRECTIONAL,
