+++ cocos/scene-graph/scene.jsb.ts
@@ -28,9 +28,12 @@ import { assert } from "../core/platform/debug";
 import { updateChildrenForDeserialize } from '../core/utils/jsb-utils';
 import { SceneGlobals } from './scene-globals';
 import { patch_cc_Scene } from '../native-binding/decorators';
+import type { Scene as JsbScene } from './scene';
 
-export const Scene = jsb.Scene;
-export type Scene = jsb.Scene;
+declare const jsb: any;
+
+export const Scene: typeof JsbScene = jsb.Scene;
+export type Scene = JsbScene;
 legacyCC.Scene = Scene;
 
 const sceneProto: any = Scene.prototype;
